Fix entrypoint for image definition (#67)
Confusingly, image definitions have an "entrypoint" field that is different from the "entrypoint" field for other definitions. Unlike, say, a Python definition, an image definition's entrypoint is not a script, but an instruction to Docker to override the default image entrypoint. All entrypoint-related functions on definitions & task kind definitions are of the script variety, & so technically image definitions don't have entrypoints. This is correct for `getEntrypoint()`, but was incorrect for `setEntrypoint()` & `{set,get}AbsoluteEntrypoint()`.